+++ examples/data-sources/kraftcloud_instance/data-source.tf
@@ -1,3 +1,3 @@
-data "kraftcloud_instance" "example" {
+data "unikraft-cloud_instance" "example" {
   uuid = "550e8400-e29b-41d4-a716-446655440000"
 }