+++ examples/provider/provider_auth.tf
@@ -1,3 +1,3 @@
-provider "kraftcloud" {
+provider "unikraft-cloud" {
   token = "cm9ib3QkbXl1c2VyLnVuaWtyYWZ0LmlvLnVzZXJzLmtyYWZ0Y2xvdWQ6a1I0ZjdFWEFNUExFS0VZ"
 }