+++ examples/resources/kraftcloud_instance/import.sh
@@ -1,1 +1,1 @@
-terraform import kraftcloud_instance.example 550e8400-e29b-41d4-a716-446655440000
+terraform import unikraft-cloud_instance.example 550e8400-e29b-41d4-a716-446655440000